pyproject.toml
[project]
name = "mcp-stocks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.52.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.2",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.0",
]
